requirements.txt
Flask
gtts
openai-whisper
Pillow
cohere
python-dotenv